[KEYCLOAK-3130] - Permission checks to authorization admin endpoints

+++ services/src/main/java/org/keycloak/authorization/admin/AuthorizationService.java
@@ -60,11 +60,15 @@ public class AuthorizationService {
     }
 
     public void enable() {
-        resourceServer().create();
+        if (!isEnabled()) {
+            resourceServer().create();
+        }
     }
 
     public void disable() {
-        resourceServer().delete();
+        if (isEnabled()) {
+            resourceServer().delete();
+        }
     }
 
     public boolean isEnabled() {
